refactor(tui): extract bg absorb and outcome dispatch from run_loop
Split the monolithic event-loop module into:
- tui/bg.rs — spawn helpers, BgChannels, absorb_background_results
- tui/dispatch.rs — KeyOutcome → IO side effects
- tui/run_loop.rs — terminal poll/draw loop only

No intentional behaviour change (issue #225).
